Access Control & Image upload finished

+++ app/Http/Controllers/PostsController.php
@@ -3,12 +3,25 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
 //To use sql query
 use DB;
 
 class PostsController extends Controller
 {
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        //guest still can see list of posts & individual posts 
+        $this->middleware('auth')->except(['index', 'show']);
+
+    }
+
     /**
      * Display a listing of the resource.
      *
@@ -58,15 +71,34 @@ class PostsController extends Controller
         //post request input validation
         $this-> validate($request, [
             'title' => 'required',
-            'body' => 'required'
+            'body' => 'required',
+            'cover_image' => 'image | mimes:jpeg,jpg,png | nullable | max:1999'
 
         ]);
+
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        } else {
+            $fileNameToStore = 'noimage.jpg';
+        }
 
         //Create Post
         $post = new Post;
         $post->title = $request->input('title');
         $post->body = $request->input('body');
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
 
         return redirect('/posts')->with('success','Post Created');
@@ -93,6 +125,11 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts/edit')->with('post', $post);
     }
 
@@ -108,14 +145,38 @@ class PostsController extends Controller
             //post request input validation
             $this-> validate($request, [
                 'title' => 'required',
-                'body' => 'required'
-
+                'body' => 'required',
+                'cover_image' => 'image | mimes:jpeg,jpg,png | nullable | max:1999'
             ]);
     
+        // Handle File Upload
+        if($request->hasFile('cover_image')){
+            // Get filename with the extension
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get just filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // Get just extension
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            // Filename to store
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+
+        }
+
             //Update Post
             $post = Post::find($id);
             $post->title = $request->input('title');
             $post->body = $request->input('body');
+
+            //if there is a cover image to update
+            if($request->hasFile('cover_image')) {
+                //delete previous image except noimage.jpg
+                if ($post->cover_image != 'noimage.jpg') {
+                    Storage::delete('public/cover_images/' . $post->cover_image);
+                }
+                $post->cover_image = $fileNameToStore;
+            }
             $post->save();
     
             return redirect('/posts')->with('success','Post Updated');
@@ -130,6 +191,17 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+
+        // Check for correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error', 'Unauthorized Page');
+        }
+        
+        if($post->cover_image != 'noimage.jpg'){
+            // Delete image
+            Storage::delete('public/cover_images/'.$post->cover_images);
+
+        }
         $post->delete();
         return redirect('/posts')->with('success','Post Removed');
     }
